refact: add routers, annotated  and new tests
