+++ fast_api/app.py
@@ -1,125 +1,14 @@
-from fastapi import Depends, FastAPI, HTTPException
-from fastapi.security import OAuth2PasswordRequestForm
-from sqlalchemy import select
-from sqlalchemy.orm import Session
+from fastapi import FastAPI
 
-from fast_api.database import get_session
-from fast_api.models import User
-from fast_api.schemas import Message, Token, UserList, UserPublic, UserSchema
-from fast_api.security import (
-    create_access_token,
-    get_current_user,
-    get_password_hash,
-    verify_password,
-)
+from routers import auth, users
 
 app = FastAPI()
 
-
-@app.post('/token', response_model=Token)
-def login_for_access_token(
-    form_data: OAuth2PasswordRequestForm = Depends(),
-    session: Session = Depends(get_session),
-):
-    user = session.scalar(select(User).where(User.email == form_data.username))
-
-    if user is None:
-        raise HTTPException(
-            status_code=400, detail='Incorrect email or password'
-        )
-
-    if not verify_password(form_data.password, user.password):
-        raise HTTPException(
-            status_code=400, detail='Incorrect email or password'
-        )
-
-    access_token = create_access_token(data={'sub': user.email})
-
-    return {'access_token': access_token, 'token_type': 'bearer'}
+app.include_router(users.router)
+app.include_router(auth.router)
 
 
 @app.get('/')
 def read_root():
     """returns a hello world message"""
     return {'message': 'Hello world! Im a root!'}
-
-
-@app.get('/users/', response_model=UserList)
-def read_users(
-    skip: int = 0, limit: int = 100, session: Session = Depends(get_session)
-):
-    """view all users registered in database"""
-    users = session.scalars(select(User).offset(skip).limit(limit)).all()
-    return {'users': users}
-
-
-@app.get('/users/{user_id}', response_model=UserPublic)
-def read_user(user_id: int, session: Session = Depends(get_session)):
-    """view only one user, by his id"""
-    db_user = session.scalar(select(User).where(User.id == user_id))
-
-    if db_user is None:
-        raise HTTPException(status_code=404, detail='User not found')
-    else:
-        return db_user
-
-
-@app.post('/users/', status_code=201, response_model=UserPublic)
-def create_user(user: UserSchema, session: Session = Depends(get_session)):
-
-    # checks if exists this user
-    db_user = session.scalar(
-        select(User).where(User.username == user.username)
-    )
-
-    if db_user:   # if exists, raise
-        raise HTTPException(
-            status_code=400, detail='Username already registered'
-        )
-
-    else:   # else, add him to database
-        hashed_password = get_password_hash(user.password)
-
-        db_user = User(
-            username=user.username, password=hashed_password, email=user.email
-        )
-        session.add(db_user)
-        session.commit()
-        session.refresh(db_user)
-
-        return db_user
-
-
-@app.put('/users/{user_id}', response_model=UserPublic)
-def update_user(
-    user_id: int,
-    user: UserSchema,
-    session: Session = Depends(get_session),
-    current_user: User = Depends(get_current_user),
-):
-    """update informations about one user, with his id"""
-
-    if current_user.id != user_id:   # checks if is logged user
-        raise HTTPException(status_code=400, detail='Not enough permissions')
-
-    current_user.username = user.username
-    current_user.password = user.password
-    current_user.email = user.email
-    return current_user
-
-
-@app.delete('/users/{user_id}', response_model=Message)
-def delete_user(
-    user_id: int,
-    session: Session = Depends(get_session),
-    current_user: User = Depends(get_current_user),
-):
-    """delete one user from the database, using his id"""
-
-    if current_user.id != user_id:
-        raise HTTPException(status_code=400, detail='Not enough permissions')
-
-    session.delete(current_user)
-    session.commit()
-
-    return {'detail': 'User deleted'}
