+++ routers/users.py
@@ -1,0 +1,90 @@
+from typing import Annotated
+
+from fastapi import APIRouter, Depends, HTTPException
+from sqlalchemy import select
+from sqlalchemy.orm import Session
+
+from fast_api.database import get_session
+from fast_api.models import User
+from fast_api.schemas import Message, UserList, UserPublic, UserSchema
+from fast_api.security import get_current_user, get_password_hash
+
+router = APIRouter(prefix='/users', tags=['users'])
+
+Session = Annotated[Session, Depends(get_session)]
+CurrentUser = Annotated[User, Depends(get_current_user)]
+
+
+@router.post('/', response_model=UserPublic, status_code=201)
+def create_user(user: UserSchema, session: Session):
+    # checks if exists this user
+    db_user = session.scalar(
+        select(User).where(User.username == user.username)
+    )
+
+    if db_user:   # if exists, raise
+        raise HTTPException(
+            status_code=400, detail='Username already registered'
+        )
+
+    else:   # else, add him to database
+        hashed_password = get_password_hash(user.password)
+
+        db_user = User(
+            username=user.username, password=hashed_password, email=user.email
+        )
+        session.add(db_user)
+        session.commit()
+        session.refresh(db_user)
+
+        return db_user
+
+
+@router.get('/', response_model=UserList)
+def read_users(session: Session, skip: int = 0, limit: int = 100):
+    """view all users registered in database"""
+    users = session.scalars(select(User).offset(skip).limit(limit)).all()
+    return {'users': users}
+
+
+@router.get('/{user_id}', response_model=UserPublic)
+def read_user(user_id: int, session: Session):
+    """view only one user, by his id"""
+    db_user = session.scalar(select(User).where(User.id == user_id))
+
+    if db_user is None:
+        raise HTTPException(status_code=404, detail='User not found')
+    else:
+        return db_user
+
+
+@router.put('/{user_id}', response_model=UserPublic)
+def update_user(
+    user_id: int, user: UserSchema, session: Session, current_user: CurrentUser
+):
+    """update informations about one user, with his id"""
+
+    if current_user.id != user_id:   # checks if is logged user
+        raise HTTPException(status_code=400, detail='Not enough permissions')
+
+    current_user.username = user.username
+    current_user.password = user.password
+    current_user.email = user.email
+    return current_user
+
+
+@router.delete('/{user_id}', response_model=Message)
+def delete_user(
+    user_id: int,
+    session: Session,
+    current_user: CurrentUser,
+):
+    """delete one user from the database, using his id"""
+
+    if current_user.id != user_id:
+        raise HTTPException(status_code=400, detail='Not enough permissions')
+
+    session.delete(current_user)
+    session.commit()
+
+    return {'detail': 'User deleted'}
